rename package owner
