+++ src/Tisstech/BootstrapLaravelScaffold/Relation.php
@@ -1,4 +1,4 @@
-<?php namespace Wfsneto\BootstrapLaravelScaffold;
+<?php namespace Tisstech\BootstrapLaravelScaffold;
 
 class Relation
 {
