+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/base-repository-interface.txt
