+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/eloquent-repository.txt
