+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/layout.txt
