+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/model-no-repo.txt
