+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/model.txt
