+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/repository-interface.txt
