+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/resource/controller.txt
