+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/resource/edit.txt
