+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/resource/test.txt
