+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/restful/controller.txt
