+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/restful/show.txt
