+++ src/Tisstech/BootstrapLaravelScaffold/templates/3-columns/restful/test.txt
