+++ src/Tisstech/BootstrapLaravelScaffold/templates/base-repository-interface.txt
