+++ src/Tisstech/BootstrapLaravelScaffold/templates/eloquent-repository.txt
