+++ src/Tisstech/BootstrapLaravelScaffold/templates/layout.txt
