+++ src/Tisstech/BootstrapLaravelScaffold/templates/model-no-repo.txt
