+++ src/Tisstech/BootstrapLaravelScaffold/templates/model.txt
