+++ src/Tisstech/BootstrapLaravelScaffold/templates/repository-interface-with-base.txt
