+++ src/Tisstech/BootstrapLaravelScaffold/templates/repository-interface.txt
