+++ src/Tisstech/BootstrapLaravelScaffold/templates/resource/controller.txt
