+++ src/Tisstech/BootstrapLaravelScaffold/templates/resource/index.txt
