+++ src/Tisstech/BootstrapLaravelScaffold/templates/resource/show.txt
