+++ src/Tisstech/BootstrapLaravelScaffold/templates/resource/test.txt
