+++ src/Tisstech/BootstrapLaravelScaffold/templates/restful/create.txt
