+++ src/Tisstech/BootstrapLaravelScaffold/templates/restful/index.txt
