+++ src/Tisstech/BootstrapLaravelScaffold/templates/restful/show.txt
