+++ src/Tisstech/BootstrapLaravelScaffold/templates/validator.txt
